[xdelta] Download and ship the xdelta3 tool.
As part of our development of the fast deployment
system we need to have a way to do binary diffs.
The xdelta tool (http://xdelta.org) provides that
capability. This PR makes use of the binaries
produced from the following fork of xdelta

	https://github.com/dellis1972/xdelta

It downloads the binaries for MacOS, Linux and Windows.
It also downloads Android binaries for the following
architectures

	x86, x86_64, armeabi-v7a, arm64-v8a

This will allow us to produce a binary diff on the
host machine and apply it on the device.